% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-fun.R
\name{has_reverse}
\alias{has_reverse}
\title{Has reverse?}
\usage{
has_reverse(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical scalar.
}
\description{
Checks whether an object has an "as_reverse" attribute.
}
