% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-fun.R
\name{is.cost_fun}
\alias{is.cost_fun}
\title{Check for type 'cost function'}
\usage{
is.cost_fun(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is a cost function.
}
