% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-fun.R
\name{is_context_sensitive}
\alias{is_context_sensitive}
\title{Is context-sensitive?}
\usage{
is_context_sensitive(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether a cost function is context-sensitive.
}
