% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-fun.R
\name{is_symmetric}
\alias{is_symmetric}
\title{Is symmetric?}
\usage{
is_symmetric(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether a cost function is symmetric.
}
