% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-fun.R
\name{is_vectorised}
\alias{is_vectorised}
\title{Is vectorised?}
\usage{
is_vectorised(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether a cost function is vectorised.
}
