% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-cost.R
\name{seq_cost}
\alias{seq_cost}
\title{Compute cost}
\usage{
seq_cost(x, ...)
}
\arguments{
\item{x}{Sequence to analyze.}

\item{...}{Further arguments to pass to \code{\link{seq_opt}}.}
}
\value{
Numeric scalar
}
\description{
Computes the cost of a given sequence.
}
